Move GLIP device gateway into library
Previously, only the osd-device-gateway used GLIP to connect to a
glip-enabled device. Connecting the two libraries, however, is a
considerable amount of work, which we would need to do over and over
again if we want to make use of the functionality from other language
bindings. Hence let's move the GLIP gateway into the library as
osd_gateway_glip() and adjust the osd-device-gateway tool to make use of
it.

The GLIP functionality is optional and gated by the --with-glip
configure option (automatically enabled if GLIP is found on the system).