Implement MAM client module
The MAM client module (osd_cl_mam) is the counterpart of the Memory
Access Module (MAM). The client speaks the MAM protocol to read and
write data to the memory, and can be used in conjunction with a host
module (which handles the communication) to transfer data between a
memory and the PC.

The MAM client module is implemented following a composition pattern to
facilitate easier testing (compared to inheritance) and to allow
multiple clients to be used at the same time to fulfill one higher-level
task (e.g. implementing a memory loader tool).