Add Python bindings using Cython
These bindings create a cpython binary extension module, which wraps the
important classes of libosd and hence provides a convenient access to
these classes from a Python environment, as if the classes were
implemented in Python itself.

Building the Python bindings requires an *installed* version of libosd
and libglip.

Limitations for now:
- Very limited testing, as most useful tests would require running OSD
  hardware. Once things stabilize a bit we can integrate for example an
  OpTiMSoC verilated compute tile to run tests.
- No Python 2 testing (hence no compatibility). Let's leave that version
  of Python behind unless strong reasons force us to do otherwise.
- It's not integrated in the build system of libosd. It's not clear if
  this is a limitation or a good thing; ideally users can get the Python
  bindings from pip more easily.