Python: Hold GIL when processing a log item
In our logging setup, the Python code gets a callback from libosd when a
log entry is to be written. We want to take this log entry and pass it
on to the logging subsystem of Python.

This results in two requirements:
1. We must hold the Python Global Interpreter Lock (GIL) when modifying
   any Python object, such calling the functions of the logging module.
2. We cannot obtain the GIL in the logging callback we get from libosd,
   as these callbacks happen in parallel with other code execution, which
   at this time blocks the main thread (and holds the GIL).

Before this patch, Requirement 1 was not followed. This caused a crash
in some cases.

If Requirement 2 is not followed we get a deadlock, as seen in this
example:
1. Call osd.Hostmod.connect(), which is forwarded to osd_hostmod_connect().
   The GIL is held by main thread
2. osd_hostmod_connect() calls (after some steps) the log callback
   log_cb() and waits for it to return before continuing its
   processing.
3. Inside the log_cb(), we want to access the Python object logging.
   To do so, we need to acquire the GIL.
This situation is a deadlock: log_cb() cannot get the GIL, and
osd_hostmod_connect() won't release it as it never returns.

In this patch, the log writing is not done directly from the log_cb()
function. Instead, the log writing request is queued and triggered by
the Python engine from the main thread as soon as it finds time to do
so.

The main limitation of this solution (or in fact, any solution) is that
log entries are not written while a call to a libosd blocks. This
limitation is inherent in the single-threaded nature of the Python
interpreter.