Re-combine split event packets upon reception
When receiving a packet of TYPE == EVENT and TYPE_SUB == EV_CONT store
this packet in a buffer and wait for a packet with the same SRC and
TYPE_SUB == EV_LAST before the packet is forwarded to the consumers.

With this logic no split packets are forwarded to the users of
osd_hostmod, only "full" event packets are seen. This makes the handling
of these packets fully transparent to the subsequent modules.