Python bindings: Make Packet modifiable
The Python-version of osd_packet is the Packet class. Up until now,
creating an OSD packet through this class resulted in a osd_packet
instance with 10 payload words.

With this commit, the Packet class becomes Python list-like object,
meaning you can use Packet like any other list. This behavior matches
the C behavior, where osd_packet is actually a simple array of uint16
words.

To access the payload of the packet we have a convenience view
Packet.payload, which again is a list. Calling Packet.payload[0] will
access for example the first payload word, and Packet.payload.append(37)
will append a payload word of value 37 to the packet.

Note that this implementation is not optimized for speed in any way. In
fact, it makes use of many memory operations which can (and in most
cases will) be inefficient. Repeatedly calling append() will cause
repeated calls to realloc() as no advanced growth strategy is
implemented. However, in most cases the C library will help to reduce
the overhead a bit by making calls to realloc() fast if adjacent memory
is available. (Which should be the case in the standard allocation
strategy that glibc uses, for example.)