Collect coverage from Python test runs
The Python bindings are generated with Cython and tested with pytest.
When running the bindings we can generate two types of coverage
information:

1. Coverage of libosd: which functionality in libosd is called through
   through the bindings, as exercised by our pytest-driven tests?
2. Coverage of the bindings itself: which functionality in the bindings
   is exercised by the pytest tests?

For (1) we rely on coverage collection functionality built into libosd,
which is enabled if the library is built with `--enable-coverage`. This
writes '*.gcda' files, compatible with gcov, whenever the library is
called. No special configuration in the Python bindings, or pytest, is
needed.

For (2) we need to configure Cython to write line coverage information
when executing code. This requires a fair bit of configuration in
various places: Cython needs to be instructed to include line coverage
information when transforming the *.pyx file into a *.c file. Then, GCC
needs to be instructed to set the define `CYTHON_TRACE_NOGIL` when
compiling the Cython-generated *.c file into a Python extension library.
All of this magic is handled by setup.py if instructed to do so.
Enabling tracing is therefore a matter of calling setup.py with
`--use-cython` (to recreate the osd.c file), and ´--cython-coverage` (to
set the appropriate configuration options). Instead of command line
arguments, the environment variables `USE_CYTHON` and `CYTHON_COVERAGE`
can be used.

All of this functionality is conveniently available through tox. Calling
`tox -e clean,coverage` uses Cython to recreate the osd.c file, compiles
the Python extension library with coverage collection enabled, runs the
pytests, and finally transforms the coverage information into a
coverage.xml file. If another form of coverage reporting is desired the
`coverage` tool can be used. (It automatically picks up the coverage
database contained in the `.coverage` file.)