Add infrastructure to run pytest with ASAN
The Python bindings contain a fair amount of tricky C code which can
benefit from ASAN. In this case, libosd must be compiled with ASAN
(`./configure --enable-asan`); the Python bindings and Python itself
don't need to be recompiled. This gives us a situation where we use a
ASAN-enabled library (libosd) in an executable (python) without ASAN. To
make that work we need to inject libasan.so with LD_PRELOAD. A new
configuration in tox and a Makefile target makes that easier.

Unfortunately, neither Ubuntu 16.04 nor 18.04 can be used to run these
tests for different reasons, see the src/python/README.md file for the
gory details.